All update

+++ CashShop/src/main/java/net/skyexcel/server/data/CashData.java
@@ -1,11 +1,11 @@
 package net.skyexcel.server.data;
 
 public interface CashData {
-    public void Increase(int Amount);
+    public void Increase(long Amount);
 
-    public void Decrease(int Amount);
+    public void Decrease(long Amount);
 
-    public void Set(int Amount);
+    public void Set(long Amount);
 
 
 }
